requirements.txt
# Core dependencies
PyYAML>=6.0
PyJWT>=2.4.0
cryptography>=3.4.8
httpx>=0.24.0
httpx[http2]

# Optional dependencies for development
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
